Selvach/code tidy up (#506)
* Refactor utils

* no message

* Rename director to snake case
