+++ src/features/assessment/DescribeSymptomsScreen.tsx
@@ -15,7 +15,7 @@ import { ValidatedTextInput } from '@covid/components/ValidatedTextInput';
 import { ValidationError } from '@covid/components/ValidationError';
 import { AssessmentInfosRequest } from '@covid/core/assessment/dto/AssessmentInfosRequest';
 import UserService, { ICoreService } from '@covid/core/user/UserService';
-import { cleanFloatVal } from '@covid/core/utils/number';
+import { cleanFloatVal } from '@covid/utils/number';
 import AssessmentCoordinator from '@covid/core/assessment/AssessmentCoordinator';
 import i18n from '@covid/locale/i18n';
 import { assessmentService } from '@covid/Services';
