+++ src/features/patient/fields/PeriodQuestion.tsx
@@ -2,7 +2,7 @@ import { FormikProps } from 'formik';
 import React, { Component } from 'react';
 
 import DropdownField from '@covid/components/DropdownField';
-import { cleanIntegerVal } from '@covid/core/utils/number';
+import { cleanIntegerVal } from '@covid/utils/number';
 import i18n from '@covid/locale/i18n';
 
 import { PeriodFrequencyQuestion, PeriodFrequencyData } from './PeriodFrequencyQuestion';
