--- src/utils/helpers.ts
@@ -1,3 +1,0 @@
-export const stripAndRound = (str: string): number => {
-  return Math.round(parseFloat(str.replace(/,/g, '')));
-};
