+++ src/utils/number.ts
@@ -1,9 +1,11 @@
-import { parse } from 'querystring';
-
 export const cleanIntegerVal = (intString: string) => {
   return parseInt(intString.replace(/ /g, ''), 10);
 };
 
 export const cleanFloatVal = (floatString: string) => {
   return parseFloat(floatString.replace(/ /g, ''));
+};
+
+export const stripAndRound = (str: string): number => {
+  return Math.round(parseFloat(str.replace(/,/g, '')));
 };
